+++ src/js/media/flash.js
@@ -132,7 +132,7 @@ vjs.Flash.prototype.setSrc = function(src){
   // e.g. Load player w/ no source, wait 2s, set src.
   if (this.player_.autoplay()) {
     var tech = this;
-    setTimeout(function(){ tech.play(); }, 0);
+    this.setTimeout(function(){ tech.play(); }, 0);
   }
 };
 
@@ -309,7 +309,7 @@ vjs.Flash['checkReady'] = function(tech){
     tech.triggerReady();
   } else {
     // wait longer
-    setTimeout(function(){
+    this.setTimeout(function(){
       vjs.Flash['checkReady'](tech);
     }, 50);
   }