+++ src/js/media/html5.js
@@ -219,7 +219,7 @@ vjs.Html5.prototype.enterFullScreen = function(){
 
     // playing and pausing synchronously during the transition to fullscreen
     // can get iOS ~6.1 devices into a play/pause loop
-    setTimeout(function(){
+    this.setTimeout(function(){
       video.pause();
       video.webkitEnterFullScreen();
     }, 0);