+++ src/js/media/media.js
@@ -177,8 +177,7 @@ vjs.MediaTechController.prototype.manualProgressOff = function(){
 };
 
 vjs.MediaTechController.prototype.trackProgress = function(){
-
-  this.progressInterval = setInterval(vjs.bind(this, function(){
+  this.progressInterval = this.setInterval(function(){
     // Don't trigger unless buffered amount is greater than last time
 
     var bufferedPercent = this.player().bufferedPercent();
@@ -192,9 +191,9 @@ vjs.MediaTechController.prototype.trackProgress = function(){
     if (bufferedPercent === 1) {
       this.stopTrackingProgress();
     }
-  }), 500);
+  }, 500);
 };
-vjs.MediaTechController.prototype.stopTrackingProgress = function(){ clearInterval(this.progressInterval); };
+vjs.MediaTechController.prototype.stopTrackingProgress = function(){ this.clearInterval(this.progressInterval); };
 
 /*! Time Tracking -------------------------------------------------------------- */
 vjs.MediaTechController.prototype.manualTimeUpdatesOn = function(){
@@ -224,14 +223,14 @@ vjs.MediaTechController.prototype.manualTimeUpdatesOff = function(){
 
 vjs.MediaTechController.prototype.trackCurrentTime = function(){
   if (this.currentTimeInterval) { this.stopTrackingCurrentTime(); }
-  this.currentTimeInterval = setInterval(vjs.bind(this, function(){
+  this.currentTimeInterval = this.setInterval(function(){
     this.player().trigger('timeupdate');
-  }), 250); // 42 = 24 fps // 250 is what Webkit uses // FF uses 15
+  }, 250); // 42 = 24 fps // 250 is what Webkit uses // FF uses 15
 };
 
 // Turn off play progress tracking (when paused or dragging)
 vjs.MediaTechController.prototype.stopTrackingCurrentTime = function(){
-  clearInterval(this.currentTimeInterval);
+  this.clearInterval(this.currentTimeInterval);
 
   // #1002 - if the video ends right before the next timeupdate would happen,
   // the progress bar won't make it all the way to the end