+++ test/unit/media.js
@@ -1,21 +1,20 @@
-var noop = function() {}, clock, featuresProgessEvents;
-
 module('Media Tech', {
   'setup': function() {
-    clock = sinon.useFakeTimers();
-    featuresProgessEvents = videojs.MediaTechController.prototype['featuresProgessEvents'];
+    this.noop = function() {};
+    this.clock = sinon.useFakeTimers();
+    this.featuresProgessEvents = videojs.MediaTechController.prototype['featuresProgessEvents'];
     videojs.MediaTechController.prototype['featuresProgressEvents'] = false;
   },
   'teardown': function() {
-    clock.restore();
-    videojs.MediaTechController.prototype['featuresProgessEvents'] = featuresProgessEvents;
+    this.clock.restore();
+    videojs.MediaTechController.prototype['featuresProgessEvents'] = this.featuresProgessEvents;
   }
 });
 
 test('should synthesize timeupdate events by default', function() {
   var timeupdates = 0, playHandler, i, tech;
   tech = new videojs.MediaTechController({
-    id: noop,
+    id: this.noop,
     on: function(event, handler) {
       if (event === 'play') {
         playHandler = handler;
@@ -32,20 +31,20 @@ test('should synthesize timeupdate events by default', function() {
     timeupdates++;
   });
 
-  clock.tick(250);
+  this.clock.tick(250);
   equal(timeupdates, 1, 'triggered one timeupdate');
 });
 
 test('stops timeupdates if the tech produces them natively', function() {
   var timeupdates = 0, tech, playHandler, expected;
   tech = new videojs.MediaTechController({
-    id: noop,
+    id: this.noop,
     on: function(event, handler) {
       if (event === 'play') {
         playHandler = handler;
       }
     },
-    bufferedPercent: noop,
+    bufferedPercent: this.noop,
     trigger: function(event) {
       if (event === 'timeupdate') {
         timeupdates++;
@@ -58,14 +57,14 @@ test('stops timeupdates if the tech produces them natively', function() {
   tech.trigger('timeupdate');
 
   expected = timeupdates;
-  clock.tick(10 * 1000);
+  this.clock.tick(10 * 1000);
   equal(timeupdates, expected, 'did not simulate timeupdates');
 });
 
 test('stops manual timeupdates while paused', function() {
   var timeupdates = 0, tech, playHandler, pauseHandler, expected;
   tech = new videojs.MediaTechController({
-    id: noop,
+    id: this.noop,
     on: function(event, handler) {
       if (event === 'play') {
         playHandler = handler;
@@ -73,7 +72,7 @@ test('stops manual timeupdates while paused', function() {
         pauseHandler = handler;
       }
     },
-    bufferedPercent: noop,
+    bufferedPercent: this.noop,
     trigger: function(event) {
       if (event === 'timeupdate') {
         timeupdates++;
@@ -81,24 +80,24 @@ test('stops manual timeupdates while paused', function() {
     }
   });
   playHandler.call(tech);
-  clock.tick(10 * 250);
+  this.clock.tick(10 * 250);
   ok(timeupdates > 0, 'timeupdates fire during playback');
 
   pauseHandler.call(tech);
   timeupdates = 0;
-  clock.tick(10 * 250);
+  this.clock.tick(10 * 250);
   equal(timeupdates, 0, 'timeupdates do not fire when paused');
 
   playHandler.call(tech);
-  clock.tick(10 * 250);
+  this.clock.tick(10 * 250);
   ok(timeupdates > 0, 'timeupdates fire when playback resumes');
 });
 
 test('should synthesize progress events by default', function() {
   var progresses = 0, tech;
   tech = new videojs.MediaTechController({
-    id: noop,
-    on: noop,
+    id: this.noop,
+    on: this.noop,
     bufferedPercent: function() {
       return 0;
     },
@@ -112,23 +111,23 @@ test('should synthesize progress events by default', function() {
     progresses++;
   });
 
-  clock.tick(500);
+  this.clock.tick(500);
   equal(progresses, 1, 'triggered one event');
 });
 
 test('dispose() should stop time tracking', function() {
   var tech = new videojs.MediaTechController({
-    id: noop,
-    on: noop,
-    off: noop,
-    trigger: noop
+    id: this.noop,
+    on: this.noop,
+    off: this.noop,
+    trigger: this.noop
   });
   tech.dispose();
 
   // progress and timeupdate events will throw exceptions after the
   // tech is disposed
   try {
-    clock.tick(10 * 1000);
+    this.clock.tick(10 * 1000);
   } catch (e) {
     return equal(e, undefined, 'threw an exception');
   }
@@ -137,8 +136,8 @@ test('dispose() should stop time tracking', function() {
 
 test('should add the source hanlder interface to a tech', function(){
   var mockPlayer = {
-    off: noop,
-    trigger: noop
+    off: this.noop,
+    trigger: this.noop
   };
   var sourceA = { src: 'foo.mp4', type: 'video/mp4' };
   var sourceB = { src: 'no-support', type: 'no-support' };